configure.in: Revert to an old implementation of "--with-xft".
We see here the very first implementation of the --with-xft configure script
option. This is the only implementation of xft within configure.in that will
actually create a working configure script on my setup (apparently this is an OS
X issue). You *could* just comment out the line ./configure complains about, but
at the expense of Xft, and I want Xft. This just works.