my xfocus patch
I added an xfocus command
   xfocus (prints status)
   xfocus 0 (turns this off)
   xfocus 1 (turns this on)

xfocus mode makes *frame* focus follow mouse.  There's no concept
of frames in X11, so doing this as an Xlib program (ie, external)
is not actually possible.  The best you can do externally is
xsloppy, which is fine if that's what you want, but I found
myself working around it a lot and being frustrated by it.

Tada.